refactored all to one file
