+++ ZachsBranch/TetrisAI/.vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\code"
+  ],
+  "SelectedNode": "\\code\\game.py",
+  "PreviewInSolutionExplorer": false
+}
